+++ carts/urls.py
@@ -1,0 +1,3 @@
+from django.urls    import path
+
+urlpatterns = []
