+++ handwash/urls.py
@@ -18,4 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
   path('users', include('users.urls')),
+  path('likes', include('likes.urls')),
+  path('carts', include('carts.urls'))
 ]
