+++ users/models.py
@@ -1,12 +1,10 @@
 from django.db       import models
-from django.db.models.fields import EmailField
 
 from core.models     import TimeStamp
-from products.models import Product, Size
 
 class User(TimeStamp):
   email        = models.CharField(max_length=200, unique=True)
-  name         = models.CharField(max_length=45)
+  name         = models.CharField(max_length=45, null=True)
   password     = models.CharField(max_length=200)
   birth        = models.DateField()
   last_name    = models.CharField(max_length=45, null=True)
@@ -19,18 +17,5 @@ class User(TimeStamp):
   class Meta:
     db_table = 'users'
 
-class Like(TimeStamp):
-  product = models.ForeignKey(Product, on_delete=models.CASCADE)
-  user    = models.ForeignKey(User, on_delete=models.CASCADE)
 
-  class Meta:
-    db_table = 'likes'
 
-class Cart(TimeStamp):
-  product = models.ForeignKey(Product, on_delete=models.CASCADE)
-  user    = models.ForeignKey(User, on_delete=models.CASCADE) 
-  size    = models.ForeignKey(Size, on_delete=models.CASCADE)
-  count   = models.IntegerField(default=1)
-
-  class Meta:
-      db_table = 'carts'
