+++ users/urls.py
@@ -1,8 +1,7 @@
 from django.urls    import path
-from users.views    import LikeView, SignUpView, SignInView
+from users.views    import SignUpView, SignInView
 
 urlpatterns = [
   path('/signup', SignUpView.as_view()),
   path('/signin', SignInView.as_view()),
-  path('/like', LikeView.as_view()),
 ]
